+++ cmd/vault/main.go
@@ -237,7 +237,7 @@ func showUsage() {
 }
 
 func showHelp() {
-	fmt.Println(`🔐 myminivault CLI v0.4.1
+	fmt.Println(`🔐 myminivault CLI v0.4.2
 Author: olelbis
 
 BASIC COMMANDS: